toi uu

+++ src/utils/delete.util.ts
@@ -1,11 +1,13 @@
 const fs = require('fs');
 
-export const deleteFile = function (file: any) {
-    fs.unlink('src/public/avatars/' + file, (err: any) => {
-        if (err) {
-            throw err;
-        }
+export const deleteFile = function (file: string | undefined): void {
+    if (file) {
+        fs.unlink('src/public/avatars/' + file, (err: any) => {
+            if (err) {
+                throw err;
+            }
 
-        console.log("Delete File successfully.");
-    });
+            console.log("Delete File successfully.");
+        });
+    }
 }
